refactor(common): dedupe DateUtil to dristi-common canonical
Adds DateUtil to PROTECTED_CLASSES so the per-module pipeline dedupes
service-local copies into dristi-common/util/. Order's and case's local
DateUtils diverged in subtle ways:

- Order had `getEPochFromLocalDate` using atStartOfDay (00:00:00).
- Case had `getEpochFromLocalDate` using LocalTime.now() (wall-clock).
- Order had `getFormattedDateFromEpoch` (date->string formatter).
- Case had `getLocalDateFromEpoch` (epoch->LocalDate).

Canonical preserves the union of all five methods so neither service's
behaviour changes. Both name spellings (EPoch vs Epoch) coexist.

Zone source moved to CommonConfiguration.getZoneId(), reading
`app.zone.id` (default Asia/Kolkata).

Changes:
- dristi-common/util/DateUtil.java: new canonical with HAND-CURATED
- dristi-common/config/CommonConfiguration.java: getZoneId() + @Value
- run_module_migration.py: PROTECTED_CLASSES += DateUtil:util
- order/internal/util/DateUtil.java: deleted (was orderDateUtil bean)
- cases/internal/util/DateUtil.java: deleted (was casesDateUtil bean)
- 5 caller import rewrites across order + cases services
- CaseServiceTest: explicit canonical import (was relying on local
  same-package wildcard before deletion)

Verify: mvn test -pl dristi-app -am -> 456+1 pass, BUILD SUCCESS
        mvn package -pl dristi-app -am -> BUILD SUCCESS
        Pipeline gates 1-8 all PASS
