+++ dristi-monolith/domain-case-lifecycle/src/main/java/org/pucar/dristi/caselifecycle/cases/internal/service/SmsNotificationService.java
@@ -9,7 +9,7 @@ import org.egov.common.contract.request.RequestInfo;
 import org.pucar.dristi.caselifecycle.cases.internal.config.Configuration;
 import org.pucar.dristi.common.kafka.Producer;
 import org.pucar.dristi.common.repository.ServiceRequestRepository;
-import org.pucar.dristi.caselifecycle.cases.internal.util.DateUtil;
+import org.pucar.dristi.common.util.DateUtil;
 import org.pucar.dristi.caselifecycle.cases.internal.web.models.SmsTemplateData;
 import org.pucar.dristi.caselifecycle.cases.internal.web.models.SMSRequest;
 import org.springframework.beans.factory.annotation.Autowired;
