--- dristi-monolith/domain-case-lifecycle/src/main/java/org/pucar/dristi/caselifecycle/cases/internal/util/DateUtil.java
@@ -1,43 +1,0 @@
-package org.pucar.dristi.caselifecycle.cases.internal.util;
-
-import lombok.extern.slf4j.Slf4j;
-import org.pucar.dristi.caselifecycle.cases.internal.config.Configuration;
-import org.springframework.stereotype.Component;
-
-import java.time.Instant;
-import java.time.LocalDate;
-import java.time.LocalTime;
-import java.time.ZoneId;
-import java.time.ZonedDateTime;
-import java.util.Date;
-
-@Slf4j
-@Component("casesDateUtil")
-public class DateUtil {
-
-    private final Configuration configuration;
-
-    public DateUtil(Configuration configuration) {
-        this.configuration = configuration;
-    }
-    /**
-     * Converts a "HH:mm:ss" time string to a Date object representing that time today
-     */
-    public Instant getInstantFrom(String time) {
-        LocalTime localTime = LocalTime.parse(time);
-        ZonedDateTime currentDate = ZonedDateTime.now(ZoneId.of(configuration.getZoneId()));
-        ZonedDateTime zonedDateTime = currentDate.with(localTime);
-
-        return Date.from(zonedDateTime.toInstant()).toInstant();
-    }
-
-    public LocalDate getLocalDateFromEpoch(long startTime) {
-        return Instant.ofEpochMilli(startTime)
-                .atZone(ZoneId.of(configuration.getZoneId()))
-                .toLocalDate();
-    }
-
-    public Long getEpochFromLocalDate(LocalDate date) {
-        return date.atTime(LocalTime.now()).atZone(ZoneId.of(configuration.getZoneId())).toInstant().toEpochMilli();
-    }
-}
