+++ dristi-monolith/domain-case-lifecycle/src/main/java/org/pucar/dristi/caselifecycle/cases/internal/util/InboxUtil.java
@@ -8,6 +8,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.egov.tracer.model.ServiceCallException;
 import org.pucar.dristi.caselifecycle.cases.internal.config.Configuration;
 import org.pucar.dristi.common.repository.ServiceRequestRepository;
+import org.pucar.dristi.common.util.DateUtil;
 import org.pucar.dristi.caselifecycle.cases.internal.web.models.OpenHearing;
 import org.pucar.dristi.caselifecycle.cases.internal.web.models.inbox.*;
 import org.springframework.stereotype.Component;
