+++ dristi-monolith/domain-case-lifecycle/src/main/java/org/pucar/dristi/caselifecycle/order/internal/service/OrderRegistrationService.java
@@ -12,7 +12,7 @@ import org.pucar.dristi.caselifecycle.order.internal.enrichment.OrderRegistratio
 import org.pucar.dristi.common.kafka.Producer;
 import org.pucar.dristi.caselifecycle.order.internal.repository.OrderRepository;
 import org.pucar.dristi.caselifecycle.order.internal.util.CaseUtil;
-import org.pucar.dristi.caselifecycle.order.internal.util.DateUtil;
+import org.pucar.dristi.common.util.DateUtil;
 import org.pucar.dristi.common.util.FileStoreUtil;
 import org.pucar.dristi.caselifecycle.order.internal.util.HearingUtil;
 import org.pucar.dristi.common.util.WorkflowUtil;
