+++ dristi-monolith/domain-case-lifecycle/src/test/java/org/pucar/dristi/caselifecycle/cases/internal/service/CaseServiceTest.java
@@ -35,6 +35,7 @@ import org.pucar.dristi.caselifecycle.cases.internal.repository.AdvocateOfficeCa
 import org.pucar.dristi.caselifecycle.cases.internal.repository.CaseRepository;
 import org.pucar.dristi.common.repository.ServiceRequestRepository;
 import org.pucar.dristi.caselifecycle.cases.internal.util.*;
+import org.pucar.dristi.common.util.DateUtil;
 import org.pucar.dristi.caselifecycle.cases.internal.validators.CaseRegistrationValidator;
 import org.pucar.dristi.caselifecycle.cases.internal.validators.EvidenceValidator;
 import org.pucar.dristi.caselifecycle.cases.internal.web.OpenApiCaseSummary;
