+++ scripts/migration/per_module/output/order_manifest.json
@@ -14,6 +14,7 @@
     "WorkflowUtil",
     "IndividualUtil",
     "ResponseInfoFactory",
+    "DateUtil",
     "KafkaProducerService",
     "Document",
     "Individual"
